+++ packages/SequenceTranslator/CHANGELOG.md
@@ -1,5 +1,17 @@
 # Sequence Translator changelog
 
+## 1.4.6 (WIP)
+
+### New features
+
+* Add PolyTool Enumerate for harmonized sequences
+
+### Bug fixes
+
+* Fix PolyTool chain, add check consistency, add tests
+* Fix PolyTool Convert tests
+* Fix the package init
+
 ## 1.4.5 (2024-10-11)
 
 ### Bug fixes